Refactor LookAtHead component with new XML config (#21)
* chore(media_view): refactor lookAtHead component with the new XML config

* fix(media_view): move lookAtHead xml component to its corresponding main dir and bump spatial sdk version

* chore(media_view): replace all references of the old component class with the new generated one

* fix(media_view): complete v0.5.5 migration for LookAtHead component

* refactor(media_view): remove unused ids file

* build: increase app version

---------
