+++ Showcases/media_view/app/src/main/assets/scenes/Composition.glxf
@@ -1,1 +1,1 @@
-{"nodes":[{"name":"privacyPopup","extras":{"meta_spatial":{"components":{"com.meta.spatial.toolkit.PanelDimensions":{"dimensions":{"value":[0.20000000298023224,0.20000000298023224],"type":"Vector2","keyString":"dimensions"}},"com.meta.spatial.toolkit.Visible":{"isVisible":{"value":false,"type":"Boolean","keyString":"isVisible"}},"com.meta.spatial.toolkit.Panel":{"panel":{"value":"@integer/5","type":"String","keyString":"panel"}},"com.meta.levinriegner.mediaview.app.immersive.component.LookAtHead":{"hasLooked":{"value":false,"type":"Boolean","keyString":"hasLooked"},"once":{"value":false,"type":"Boolean","keyString":"once"},"zOffset":{"value":0.7,"type":"Float","keyString":"zOffset"}}},"version":1}}},{"name":"whatsNew","translation":[0,-0.004070401191711426,0],"extras":{"meta_spatial":{"components":{"com.meta.spatial.toolkit.Visible":{"isVisible":{"value":false,"type":"Boolean","keyString":"isVisible"}},"com.meta.spatial.toolkit.PanelDimensions":{"dimensions":{"value":[0.8999999761581421,0.44999998807907104],"type":"Vector2","keyString":"dimensions"}},"com.meta.spatial.toolkit.Panel":{"panel":{"value":"@integer/7","type":"String","keyString":"panel"}}},"version":1}}},{"name":"mediaTypesFilter","translation":[0.49000000953674316,0,0],"extras":{"meta_spatial":{"components":{"com.meta.spatial.toolkit.Visible":{"isVisible":{"value":false,"type":"Boolean","keyString":"isVisible"}},"com.meta.spatial.toolkit.PanelDimensions":{"dimensions":{"value":[0.17499999701976776,0.49000000953674316],"type":"Vector2","keyString":"dimensions"}},"com.meta.spatial.toolkit.Panel":{"panel":{"value":"@integer/3","type":"String","keyString":"panel"}}},"version":1}}},{"name":"onboarding","extras":{"meta_spatial":{"components":{"com.meta.spatial.toolkit.Visible":{"isVisible":{"value":false,"type":"Boolean","keyString":"isVisible"}},"com.meta.spatial.toolkit.PanelDimensions":{"dimensions":{"value":[0.8999999761581421,0.44999998807907104],"type":"Vector2","keyString":"dimensions"}},"com.meta.spatial.toolkit.Panel":{"panel":{"value":"@integer/6","type":"String","keyString":"panel"}}},"version":1}}},{"name":"galleryMenu","translation":[0,0.2977507710456848,0],"extras":{"meta_spatial":{"components":{"com.meta.spatial.toolkit.Visible":{"isVisible":{"value":false,"type":"Boolean","keyString":"isVisible"}},"com.meta.spatial.toolkit.PanelDimensions":{"dimensions":{"value":[0.5,0.10000000149011612],"type":"Vector2","keyString":"dimensions"}},"com.meta.spatial.toolkit.Panel":{"panel":{"value":"@integer/2","type":"String","keyString":"panel"}}},"version":1}}},{"name":"gallery","children":[3,1,2,4],"extras":{"meta_spatial":{"components":{"com.meta.spatial.toolkit.PanelDimensions":{"dimensions":{"value":[0.7799999713897705,0.49000000953674316],"type":"Vector2","keyString":"dimensions"}},"com.meta.spatial.toolkit.Grabbable":{},"com.meta.spatial.toolkit.Visible":{"isVisible":{"value":false,"type":"Boolean","keyString":"isVisible"}},"com.meta.spatial.toolkit.Panel":{"panel":{"value":"@integer/1","type":"String","keyString":"panel"}},"com.meta.levinriegner.mediaview.app.immersive.component.LookAtHead":{"hasLooked":{"value":false,"type":"Boolean","keyString":"hasLooked"},"once":{"value":true,"type":"Boolean","keyString":"once"},"zOffset":{"value":1.1,"type":"Float","keyString":"zOffset"}}},"version":1}}}],"scenes":[{"name":"","nodes":[5,0]}],"scene":0,"asset":{"experience":false,"copyright":"","version":"2.0","generator":"","minVersion":"2.0"}}
+{"nodes":[{"name":"privacyPopup","extras":{"meta_spatial":{"components":{"com.meta.spatial.toolkit.PanelDimensions":{"dimensions":{"value":[0.20000000298023224,0.20000000298023224],"type":"Vector2","keyString":"dimensions"}},"com.meta.spatial.toolkit.Visible":{"isVisible":{"value":false,"type":"Boolean","keyString":"isVisible"}},"com.meta.spatial.toolkit.Panel":{"panel":{"value":"@integer/5","type":"String","keyString":"panel"}},"com.meta.levinriegner.mediaview.LookAtHead":{"hasLooked":{"value":false,"type":"Boolean","keyString":"hasLooked"},"once":{"value":false,"type":"Boolean","keyString":"once"},"zOffset":{"value":0.7,"type":"Float","keyString":"zOffset"}}},"version":1}}},{"name":"whatsNew","translation":[0,-0.004070401191711426,0],"extras":{"meta_spatial":{"components":{"com.meta.spatial.toolkit.Visible":{"isVisible":{"value":false,"type":"Boolean","keyString":"isVisible"}},"com.meta.spatial.toolkit.PanelDimensions":{"dimensions":{"value":[0.8999999761581421,0.44999998807907104],"type":"Vector2","keyString":"dimensions"}},"com.meta.spatial.toolkit.Panel":{"panel":{"value":"@integer/7","type":"String","keyString":"panel"}}},"version":1}}},{"name":"mediaTypesFilter","translation":[0.49000000953674316,0,0],"extras":{"meta_spatial":{"components":{"com.meta.spatial.toolkit.Visible":{"isVisible":{"value":false,"type":"Boolean","keyString":"isVisible"}},"com.meta.spatial.toolkit.PanelDimensions":{"dimensions":{"value":[0.17499999701976776,0.49000000953674316],"type":"Vector2","keyString":"dimensions"}},"com.meta.spatial.toolkit.Panel":{"panel":{"value":"@integer/3","type":"String","keyString":"panel"}}},"version":1}}},{"name":"onboarding","extras":{"meta_spatial":{"components":{"com.meta.spatial.toolkit.Visible":{"isVisible":{"value":false,"type":"Boolean","keyString":"isVisible"}},"com.meta.spatial.toolkit.PanelDimensions":{"dimensions":{"value":[0.8999999761581421,0.44999998807907104],"type":"Vector2","keyString":"dimensions"}},"com.meta.spatial.toolkit.Panel":{"panel":{"value":"@integer/6","type":"String","keyString":"panel"}}},"version":1}}},{"name":"galleryMenu","translation":[0,0.2977507710456848,0],"extras":{"meta_spatial":{"components":{"com.meta.spatial.toolkit.Visible":{"isVisible":{"value":false,"type":"Boolean","keyString":"isVisible"}},"com.meta.spatial.toolkit.PanelDimensions":{"dimensions":{"value":[0.5,0.10000000149011612],"type":"Vector2","keyString":"dimensions"}},"com.meta.spatial.toolkit.Panel":{"panel":{"value":"@integer/2","type":"String","keyString":"panel"}}},"version":1}}},{"name":"gallery","children":[3,1,2,4],"extras":{"meta_spatial":{"components":{"com.meta.spatial.toolkit.PanelDimensions":{"dimensions":{"value":[0.7799999713897705,0.49000000953674316],"type":"Vector2","keyString":"dimensions"}},"com.meta.spatial.toolkit.Grabbable":{},"com.meta.spatial.toolkit.Visible":{"isVisible":{"value":false,"type":"Boolean","keyString":"isVisible"}},"com.meta.spatial.toolkit.Panel":{"panel":{"value":"@integer/1","type":"String","keyString":"panel"}},"com.meta.levinriegner.mediaview.LookAtHead":{"hasLooked":{"value":false,"type":"Boolean","keyString":"hasLooked"},"once":{"value":true,"type":"Boolean","keyString":"once"},"zOffset":{"value":1.1,"type":"Float","keyString":"zOffset"}}},"version":1}}}],"scenes":[{"name":"","nodes":[5,0]}],"scene":0,"asset":{"experience":false,"copyright":"","version":"2.0","generator":"","minVersion":"2.0"}}
