+++ Showcases/media_view/app/src/main/assets/scenes/Composition/Main.metaspatialcomp
@@ -9,4 +9,4 @@ entities:
           scenes:
             - ref:Scene
 metadata:
-  version: 1.0
+  version: 1.29
