+++ Showcases/media_view/app/src/main/assets/scenes/Main.metaspatial
@@ -1,3 +1,7 @@
-entities: ~
+entities:
+  com.meta.models.TargetModel:
+    - components:
+        com.meta.components.TargetComponent:
+          platform: horizon_os
 metadata:
-  version: 1.0
+  version: 1.29
