+++ Showcases/media_view/app/src/main/assets/scenes/config.json
@@ -1,0 +1,6 @@
+{
+    "spatial.editor.customComponentXmlsPath": [
+        "../src/main/components/"
+    ],
+    "spatial.editor.libraryComponentXmlsPath": "../build/generated/components"
+}
