+++ Showcases/media_view/app/src/main/components/LookAtHead.xml
@@ -1,0 +1,11 @@
+<!--
+~ (c) Meta Platforms, Inc. and affiliates. Confidential and proprietary.
+  -->
+
+<ComponentSchema packageName="com.meta.levinriegner.mediaview" >
+    <Component name="LookAtHead" description="A component to rotate an object to look at the user's head">
+        <FloatAttribute name="zOffset" defaultValue="1f" />
+        <BooleanAttribute name="once" defaultValue="true" />
+        <BooleanAttribute name="hasLooked" defaultValue="false" />
+    </Component>
+</ComponentSchema>
