+++ Showcases/media_view/build.gradle.kts
@@ -2,7 +2,7 @@
 
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
-  id("com.android.application") version "8.1.4" apply false
+  id("com.android.application") version "8.6.1" apply false
   id("org.jetbrains.kotlin.android") version "1.9.25" apply false
   id("com.google.dagger.hilt.android") version "2.51" apply false
   id("com.datadoghq.dd-sdk-android-gradle-plugin") version "1.14.0" apply false
